+++ CHANGELOG.md
@@ -11,6 +11,14 @@ before starting to add changes. Use example [placed in the end of the page](#exa
 
 ## [Unreleased]
 
+- [PR-333](https://github.com/OS2Forms/os2forms/pull/333)
+  Added the ability to configure on which submission states handlers should run.
+  The default option is to run on the completed state. Changes was made to the
+  following handlers:
+  - Digital post
+  - Fasit
+  - FBS
+
 ## [5.1.0] 2026-06-03
 
 - [PR-326](https://github.com/OS2Forms/os2forms/pull/326)